+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Revision history for HsYAML-aeson
 
+## 0.2.0.0
+
+* Changes in accordance with HsYAML-0.2.0.0 (#1)
+    * New convinence function `decode1Strict`
+    * New functions`encode1`, `encode1Strict`, `encodeValue`, `encodeValue'` for encoding JSON Values
+
 ## 0.1.0.0
 
 * First release. Released on an unsuspecting world.